bug fixes - destroying blocks

+++ src/common/mining-pools/pool/pool-management/pool-data/block-informations/Pool-Data-Block-Information-Miner-Instance.js
@@ -36,7 +36,9 @@ class PoolDataBlockInformationMinerInstance {
         this.blockInformation = undefined;
         this.minerInstance = undefined;
 
-        this.workBlock.destroyBlock();
+        if (this.workBlock !== undefined)
+            this.workBlock.destroyBlock();
+
         this.workBlock = undefined;
 
     }
