+++ src/common/mining-pools/pool/pool-management/pool-data/block-informations/Pool-Data-Block-Information.js
@@ -49,7 +49,9 @@ class PoolDataBlockInformation {
 
         this.blockInformationMinersInstances = [];
 
-        this.block.destroyBlock();
+        if (this.block !== undefined)
+            this.block.destroyBlock();
+
         this.block = undefined;
 
     }
